Create $UDOG
What is $UDOG? $UDOG is a cryptocurrency we launched 48 hours ago
on the Blum platform. $UDOG is inspired by the Telegram symbol. After
Pavel Durov sold VKontakte, whose symbol was Spotty the dog,
Pavel made it the Telegram symbol as a sign of respect, but it returned with a new look
and the name "Resistance Dog" in honor of the fight for
freedom of speech. After which, this dog truly became a cult in the Telegram community.
The community liked the dog so much that it became legendary and a symbol of freedom
on Telegram. Then, on September 28, the first limited-edition collectible gifts were
released on Telegram, which could be upgraded and converted into NFT tokens.
Collectible gifts began breaking records, reaching their highest on June 1. The price for a
regular Pepe NFT toy started at $17,000.
The frogs themselves were released into the gift shop on January 1st.
After June 1, interest peaked. Each subsequent gift sale sold out within seconds. For
example, the 815,000 gifts for Independence Day sold out in just two minutes. And
everyone was waiting for just one gift: the legendary Resistance Dog.
On August 14th, in honor of Telegram's birthday, the long-awaited dog finally appeared in
the gift shop. Its price was 5,000 stars of Telegram's internal currency, or 22 TON. The
very next day, the unupgraded dog was selling on the secondary market for 160 TON. The
minimum and maximum purchase price was 240 TON per piece. Currently, the dog is
truly legendary; people gather in communities and discuss the growth of their gift. That's
why we chose Redo. (short for resistance dog) for our token. Its purpose is to unite all
users who want to promote the idea of raising the price of the dog. Redo has long been
more than just a dog—it's a Telegram symbol.